reconstructing project zone
