+++ src/App.tsx
@@ -2,30 +2,18 @@ import React, { useEffect, useState } from "react";
 import { Collapse } from "@mui/material";
 import Box from "@mui/material/Box";
 import styled from "styled-components";
-import BasicTabs from "./components/BasicTabs";
-import BasicTabsNoHidden from "./components/BasicTabsNoHidden";
-import { Button } from "@mui/material";
 import * as initialObj from "./module/initialObj";
-import "./App.css";
-import FolderList from "./components/FolderList";
-import { width } from "@mui/system";
 import getAPI from "./module/getAPI";
 import TopToolBarAria from "./TopToolBarAria/TopToolBarAria";
 import LeftZone from "./LeftZone/LeftZone";
-import FileList from "./components/FileList";
-import { useSelector, useDispatch } from "react-redux";
+import ProjectZone from "./ProjectZone/ProjectZone";
+
+import { useSelector } from "react-redux";
 import {
   selectViewStatsLeft,
   selectViewStatsMidUpper,
   selectViewStatsRight,
 } from "./redux/slice/ViewStatsSlice";
-import {
-  changeShouldShowDirPathInMidMain,
-  addOpeningDirInTree,
-  deleteOpeningDirInTree,
-  selectShouldShowDirPathInMidMain,
-  selectOpeningDirInTree,
-} from "./redux/slice/SelectedFolderStatsSlice";
 
 const FlexParent = styled.div`
   display: flex;
@@ -146,16 +134,7 @@ function App() {
   const ViewStatsLeft = useSelector(selectViewStatsLeft);
   const ViewStatsMidUpper = useSelector(selectViewStatsMidUpper);
   const ViewStatsRight = useSelector(selectViewStatsRight);
-  const OpeningDirInTree = useSelector(selectOpeningDirInTree);
-  const dispatch = useDispatch();
 
-  const isOpenFunc = (value: string) => {
-    return OpeningDirInTree.includes(value);
-  };
-
-  const [mainWindowStats, setMainWindowStats] = useState(
-    initialObj.mainWindowStats
-  );
   const [config, setConfig] = useState<api.Config>(initialObj.config);
   useEffect(() => {
     getAPI("get_lanch_config", setConfig);
@@ -220,24 +199,7 @@ function App() {
                       height: "100%",
                     }}
                   >
-                    <BasicTabs labels={["Files", "DAW View", "Spectrogram"]}>
-                      <div>
-                        <FileList
-                          dirPath={
-                            mainWindowStats.selectedFolderStats
-                              .shouldShowDirPathInMidMain
-                          }
-                        />
-                      </div>
-                      <Box
-                        sx={{
-                          width: "100%",
-                          height: "100%",
-                        }}
-                      >
-                        test
-                      </Box>
-                    </BasicTabs>
+                    <ProjectZone />
                   </Box>
                 </FlexMainZoneProjectZone>
               </FlexMainZoneBox>
